perf(profiling): split memalloc heap tracking into inline fast path and cold slow path
Move the allocation sampling decision (counter increment + threshold
compare) into an always-inlined function in the header so it compiles
directly into callers. The expensive work (reentrancy guard, GC guard,
traceback collection, pprof export) stays in a separate slow-path
function in the .cpp, called only when a sample is due (~0.01% of
allocations).

This avoids a cross-TU function call on every Python object allocation
while keeping the cold path out of the instruction cache.
